requirements.txt
streamlit==1.14.0
newspaper3k==0.2.8
requests==2.28.1
urllib==1.26.7
beautifulsoup4==4.11.1
nltk==3.7